+++ contract-common/src/test/groovy/org/seekay/contract/common/match/body/SymmetricJsonBodyMatcherSpec.groovy
@@ -7,10 +7,10 @@ import spock.lang.Specification
 
 import static org.seekay.contract.model.ContractTestFixtures.*
 
-class JsonBodyMatcherSpec extends Specification {
+class SymmetricJsonBodyMatcherSpec extends Specification {
 
     ObjectMapper objectMapper = new ObjectMapper()
-    JsonBodyMatcher matcher = new JsonBodyMatcher()
+    SymmetricJsonBodyMatcher matcher = new SymmetricJsonBodyMatcher()
     ExpressionMatcher expressionMatcher = Mock(ExpressionMatcher)
 
     def setup() {
@@ -44,14 +44,14 @@ class JsonBodyMatcherSpec extends Specification {
             matcher.isMatch(contractBody, actualBody)
     }
 
-    def 'a json body that is a subset of another should match' () {
+    def 'a json body that is a subset of another should not match' () {
         given:
             def contract = ["one":"two"]
             def actual = ["one":"two", "three":"four"]
             String contractBody = objectMapper.writeValueAsString(contract)
             String actualBody = objectMapper.writeValueAsString(actual)
         expect:
-            matcher.isMatch(contractBody, actualBody)
+            !matcher.isMatch(contractBody, actualBody)
     }
 
     def 'a json body where keys have the same name but different types should not match' () {
@@ -127,7 +127,7 @@ class JsonBodyMatcherSpec extends Specification {
     def 'an error thrown parsing json should be treated as not a match' () {
         given:
             ObjectMapper mockObjectMapper = Mock(ObjectMapper)
-            JsonBodyMatcher matcher = new JsonBodyMatcher(objectMapper: mockObjectMapper)
+            SymmetricJsonBodyMatcher matcher = new SymmetricJsonBodyMatcher(objectMapper: mockObjectMapper)
             def contract = ["one":["two","three","four","five"]]
             def actual = ["one":["five","four","three"]]
             String contractBody = this.objectMapper.writeValueAsString(contract)
@@ -154,39 +154,31 @@ class JsonBodyMatcherSpec extends Specification {
 
     def 'out of order elements should match' () {
         given:
-            BodyMatchingService jsonBodyMatcher = ApplicationContext.bodyMatchingService()
+            SymmetricJsonBodyMatcher symmetricJsonBodyMatcher = ApplicationContext.symmetricJsonBodyMatcher()
             String contract = '''{"url":"https://github.com/harmingcola/kvServerContracts.git","name":"kvServerContracts"}'''
             String actual = '''{"name":"kvServerContracts","url":"https://github.com/harmingcola/kvServerContracts.git"}'''
         when:
-            def match = jsonBodyMatcher.isMatch(contract, actual)
+            def match = symmetricJsonBodyMatcher.isMatch(contract, actual)
         then:
             match
     }
 
-    def 'checking' () {
-        given:
-            ObjectMapper objectMapper = new ObjectMapper()
-            JsonBodyMatcher matcher = new JsonBodyMatcher()
-            ExpressionMatcher expressionMatcher = new ExpressionMatcher()
-            matcher.objectMapper = objectMapper
-            matcher.expressionMatcher = expressionMatcher
-            String contract = '''{"value" : "${contract.anyNumber}" }'''
-            String actual = '''{"value" : 24 }'''
+    def 'an exception thrown during processing should result in no match' () {
         when:
-            def isMatch = matcher.isMatch(contract, actual)
+            def result = matcher.isMatch("one", "one")
         then:
-            isMatch
+            objectMapper.readValue(_ as String, Object.class) >> {throw new IOException()}
+            result == false
     }
 
-
-    def 'From failed test 001' () {
+    def 'Failure 001' () {
         given:
-            JsonBodyMatcher jsonBodyMatcher = ApplicationContext.jsonBodyMatcher()
-            String contract = '''{"id":"${contract.var.string.repoId}","name":"kvServerContracts","url":"https://github.com/harmingcola/kvServerContracts.git"}'''
-            String actual = '''{"id":"b03da0f8-e98e-4c9e-a609-d5d1fb2eb6d5","url":"https://github.com/harmingcola/kvServerContracts.git","name":"kvServerContracts","username":null,"password":null}'''
-        expect:
-            jsonBodyMatcher.isMatch(contract, actual)
+            SymmetricJsonBodyMatcher symmetricJsonBodyMatcher = ApplicationContext.symmetricJsonBodyMatcher()
+            String contract = '''{"name":"${contract.var.string.repoName}","url":"${contract.var.string.repoUrl}"}'''
+            String actual = '''{"url":"https://github.com/harmingcola/kvServerContracts.git","name":"Test Repo"}'''
+        when:
+            def match = symmetricJsonBodyMatcher.isMatch(contract, actual)
+        then:
+            match
     }
-
-
 }