+++ contract-model/src/main/java/org/seekay/contract/model/domain/Method.java
@@ -2,6 +2,6 @@ package org.seekay.contract.model.domain;
 
 public enum Method {
 
-    GET, POST, PUT, DELETE, CONNECT
+    GET, POST, PUT, DELETE, HEAD, OPTIONS, CONNECT
 
 }