+++ contract-server/src/main/java/org/seekay/contract/server/servet/RequestHandlerServlet.java
@@ -20,64 +20,85 @@ import static org.seekay.contract.server.util.ResponseWriter.to;
 
 public class RequestHandlerServlet extends HttpServlet {
 
-    private MatchingService matchingService;
-    private EnricherService enricherService;
-    private ContractService contractService;
+  private MatchingService matchingService;
+  private EnricherService enricherService;
+  private ContractService contractService;
 
-    @Override
-    public void init() throws ServletException {
-        matchingService = matchingService();
-        enricherService = enricherService();
-        contractService = contractService();
-    }
+  @Override
+  public void init() throws ServletException {
+    matchingService = matchingService();
+    enricherService = enricherService();
+    contractService = contractService();
+  }
 
-    @Override
-    protected void doGet(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
-        ContractRequest contractRequest = from(httpRequest).toContractRequest();
-        Contract contract = matchingService.matchGetRequest(contractRequest);
-        enableFilters(contract);
-        contract = enricherService.enrichResponse(contract);
-        writeResponse(httpsResponse, contract);
-    }
+  @Override
+  protected void doGet(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchGetRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpsResponse, contract);
+  }
 
-    @Override
-    protected void doPost(HttpServletRequest httpRequest, HttpServletResponse httpResponse) throws ServletException, IOException {
-        ContractRequest contractRequest = from(httpRequest).toContractRequest();
-        Contract contract = matchingService.matchPostRequest(contractRequest);
-        contract = enricherService.enrichResponse(contract);
-        writeResponse(httpResponse, contract);
-    }
+  @Override
+  protected void doPost(HttpServletRequest httpRequest, HttpServletResponse httpResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchPostRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpResponse, contract);
+  }
 
-    @Override
-    protected void doPut(HttpServletRequest httpRequest, HttpServletResponse httpResponse) throws ServletException, IOException {
-        ContractRequest contractRequest = from(httpRequest).toContractRequest();
-        Contract contract = matchingService.matchPutRequest(contractRequest);
-        contract = enricherService.enrichResponse(contract);
-        writeResponse(httpResponse, contract);
-    }
+  @Override
+  protected void doPut(HttpServletRequest httpRequest, HttpServletResponse httpResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchPutRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpResponse, contract);
+  }
 
-    @Override
-    protected void doDelete(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
-        ContractRequest contractRequest = from(httpRequest).toContractRequest();
-        Contract contract = matchingService.matchDeleteRequest(contractRequest);
-        contract = enricherService.enrichResponse(contract);
-        writeResponse(httpsResponse, contract);
-    }
+  @Override
+  protected void doDelete(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchDeleteRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpsResponse, contract);
+  }
 
-    private void writeResponse(HttpServletResponse httpResponse, Contract contract) throws JsonProcessingException {
-        if (contract != null && contract.getResponse() != null) {
-            to(httpResponse)
-                    .status(contract.getResponse().getStatus())
-                    .headers(contract.getResponse().getHeaders())
-                    .write(contract.getResponse().getBody());
-        } else {
-            to(httpResponse).notFound();
-        }
-    }
+  @Override
+  protected void doHead(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchHeadRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpsResponse, contract);
+  }
 
-    private void enableFilters(Contract contract) {
-        if(contract != null) {
-            contractService.enableFilters(contract.getFilters());
-        }
+  @Override
+  protected void doOptions(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
+    ContractRequest contractRequest = from(httpRequest).toContractRequest();
+    Contract contract = matchingService.matchOptionsRequest(contractRequest);
+    enableFilters(contract);
+    contract = enricherService.enrichResponse(contract);
+    writeResponse(httpsResponse, contract);
+  }
+
+  private void writeResponse(HttpServletResponse httpResponse, Contract contract) throws JsonProcessingException {
+    if (contract != null && contract.getResponse() != null) {
+      to(httpResponse)
+          .status(contract.getResponse().getStatus())
+          .headers(contract.getResponse().getHeaders())
+          .write(contract.getResponse().getBody());
+    } else {
+      to(httpResponse).notFound();
     }
+  }
+
+  private void enableFilters(Contract contract) {
+    if (contract != null) {
+      contractService.enableFilters(contract.getFilters());
+    }
+  }
 }